requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Pydantic==0.11.0
flask-restful==0.3.9
google-cloud-logging==3.5.0
gunicorn==20.1.0
pydantic==1.10.7
pytest==7.3.1
requests==2.28.2
google-cloud-bigquery==3.10.0
